+++ memory-db/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2017, 2018 Parity Technologies
+// Copyright 2017-2020 Parity Technologies
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,9 +19,12 @@
 #[cfg(not(feature = "std"))]
 extern crate alloc;
 
+mod malloc_size_of;
+pub use malloc_size_of::*;
+
 use hash_db::{HashDB, HashDBRef, PlainDB, PlainDBRef, Hasher as KeyHasher,
 	AsHashDB, AsPlainDB, Prefix};
-use parity_util_mem::{MallocSizeOf, MallocSizeOfOps};
+use parity_util_mem::{MallocSizeOf, MallocSizeOfOps, MallocShallowSizeOf};
 #[cfg(feature = "deprecated")]
 #[cfg(feature = "std")]
 use heapsize::HeapSizeOf;
@@ -70,6 +73,9 @@ impl<T> MaybeDebug for T {}
 /// the data with `get()`. Clear with `clear()` and purge the portions of the data
 /// that have no references with `purge()`.
 ///
+/// If you're not using the `MallocSizeOf` implementation to track memory usage,
+/// set the `M` type parameter to `NoopTracker`.
+///
 /// # Example
 /// ```rust
 ///
@@ -107,36 +113,48 @@ impl<T> MaybeDebug for T {}
 ///   assert!(!m.contains(&k, EMPTY_PREFIX));
 /// }
 /// ```
-pub struct MemoryDB<H, KF, T>
-	where
+pub struct MemoryDB<H, KF, T, M = MemCounter<T>>
+where
 	H: KeyHasher,
 	KF: KeyFunction<H>,
+	M: MemTracker<T>,
 {
 	data: HashMap<KF::Key, (T, i32)>,
+	// We cache `size_of(data) - shallow_size_of(data)` to compute
+	// `size_of(data)` incrementally and avoid iterating over the `data`.
+	malloc_tracker: M,
 	hashed_null_node: H::Out,
 	null_node_data: T,
 	_kf: PhantomData<KF>,
 }
 
-impl<H: KeyHasher, KF: KeyFunction<H>, T: Clone> Clone for MemoryDB<H, KF, T> {
+impl<H, KF, T, M> Clone for MemoryDB<H, KF, T, M>
+where
+	H: KeyHasher,
+	KF: KeyFunction<H>,
+	T: Clone,
+	M: MemTracker<T> + Copy,
+{
 	fn clone(&self) -> Self {
 		Self {
 			data: self.data.clone(),
 			hashed_null_node: self.hashed_null_node,
 			null_node_data: self.null_node_data.clone(),
+			malloc_tracker: self.malloc_tracker,
 			_kf: Default::default(),
 		}
 	}
 }
 
-impl<H, KF, T> PartialEq<MemoryDB<H, KF, T>> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> PartialEq<MemoryDB<H, KF, T, M>> for MemoryDB<H, KF, T, M>
 	where
 	H: KeyHasher,
 	KF: KeyFunction<H>,
 	<KF as KeyFunction<H>>::Key: Eq + MaybeDebug,
 	T: Eq + MaybeDebug,
+	M: MemTracker<T> + PartialEq,
 {
-	fn eq(&self, other: &MemoryDB<H, KF, T>) -> bool {
+	fn eq(&self, other: &MemoryDB<H, KF, T, M>) -> bool {
 		for a in self.data.iter() {
 			match other.data.get(&a.0) {
 				Some(v) if v != a.1 => return false,
@@ -148,12 +166,13 @@ impl<H, KF, T> PartialEq<MemoryDB<H, KF, T>> for MemoryDB<H, KF, T>
 	}
 }
 
-impl<H, KF, T> Eq for MemoryDB<H, KF, T>
+impl<H, KF, T, M> Eq for MemoryDB<H, KF, T, M>
 	where
 		H: KeyHasher,
 		KF: KeyFunction<H>,
 		<KF as KeyFunction<H>>::Key: Eq + MaybeDebug,
 		T: Eq + MaybeDebug,
+		M: MemTracker<T> + Eq,
 {}
 
 pub trait KeyFunction<H: KeyHasher> {
@@ -224,13 +243,15 @@ pub fn prefixed_key<H: KeyHasher>(key: &H::Out, prefix: Prefix) -> Vec<u8> {
 	prefixed_key
 }
 
-#[derive(Clone, Debug)]
 /// Key function that concatenates prefix and hash.
 /// This is doing useless computation and should only be
 /// used for legacy purpose.
 /// It shall be remove in the future.
+#[derive(Clone, Debug)]
+#[deprecated(since="0.22.0")]
 pub struct LegacyPrefixedKey<H: KeyHasher>(PhantomData<H>);
 
+#[allow(deprecated)]
 impl<H: KeyHasher> KeyFunction<H> for LegacyPrefixedKey<H> {
 	type Key = Vec<u8>;
 
@@ -241,6 +262,7 @@ impl<H: KeyHasher> KeyFunction<H> for LegacyPrefixedKey<H> {
 
 /// Legacy method for db using previous version of prefix encoding.
 /// Only for trie radix 16 trie.
+#[deprecated(since="0.22.0")]
 pub fn legacy_prefixed_key<H: KeyHasher>(key: &H::Out, prefix: Prefix) -> Vec<u8> {
 	let mut prefixed_key = Vec::with_capacity(key.as_ref().len() + prefix.0.len() + 1);
 	if let Some(last) = prefix.1 {
@@ -258,11 +280,12 @@ pub fn legacy_prefixed_key<H: KeyHasher>(key: &H::Out, prefix: Prefix) -> Vec<u8
 	prefixed_key
 }
 
-impl<'a, H, KF, T> Default for MemoryDB<H, KF, T>
+impl<'a, H, KF, T, M> Default for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: From<&'a [u8]>,
 	KF: KeyFunction<H>,
+	M: MemTracker<T> + Default,
 {
 	fn default() -> Self {
 		Self::from_null_node(&[0u8][..], [0u8][..].into())
@@ -270,11 +293,12 @@ where
 }
 
 /// Create a new `MemoryDB` from a given null key/data
-impl<H, KF, T> MemoryDB<H, KF, T>
+impl<H, KF, T, M> MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default,
 	KF: KeyFunction<H>,
+	M: MemTracker<T>,
 {
 	/// Remove an element and delete it from storage if reference count reaches zero.
 	/// If the value was purged, return the old value.
@@ -286,24 +310,29 @@ where
 		match self.data.entry(key) {
 			Entry::Occupied(mut entry) =>
 				if entry.get().1 == 1 {
-					Some(entry.remove().0)
+					let (value, _) = entry.remove();
+					self.malloc_tracker.on_remove(&value);
+					Some(value)
 				} else {
 					entry.get_mut().1 -= 1;
 					None
 				},
 			Entry::Vacant(entry) => {
-				entry.insert((T::default(), -1)); // FIXME: shouldn't it be purged?
+				let value = T::default();
+				self.malloc_tracker.on_insert(&value);
+				entry.insert((value, -1));
 				None
 			}
 		}
 	}
 }
 
-impl<'a, H: KeyHasher, KF, T> MemoryDB<H, KF, T>
+impl<'a, H, KF, T, M> MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: From<&'a [u8]>,
 	KF: KeyFunction<H>,
+	M: MemTracker<T> + Default,
 {
 	/// Create a new `MemoryDB` from a given null key/data
 	pub fn from_null_node(null_key: &'a [u8], null_node_data: T) -> Self {
@@ -311,6 +340,7 @@ where
 			data: HashMap::default(),
 			hashed_null_node: H::hash(null_key),
 			null_node_data,
+			malloc_tracker: M::default(),
 			_kf: Default::default(),
 		}
 	}
@@ -350,16 +380,25 @@ where
 	/// }
 	/// ```
 	pub fn clear(&mut self) {
+		self.malloc_tracker.on_clear();
 		self.data.clear();
 	}
 
 	/// Purge all zero-referenced data from the database.
 	pub fn purge(&mut self) {
-		self.data.retain(|_, &mut (_, rc)| rc != 0);
+		let malloc_tracker = &mut self.malloc_tracker;
+		self.data.retain(|_, (v, rc)| {
+			let keep = *rc != 0;
+			if !keep {
+				malloc_tracker.on_remove(v);
+			}
+			keep
+		});
 	}
 
-	/// Return the internal map of hashes to data, clearing the current state.
+	/// Return the internal key-value HashMap, clearing the current state.
 	pub fn drain(&mut self) -> HashMap<KF::Key, (T, i32)> {
+		self.malloc_tracker.on_clear();
 		mem::replace(&mut self.data, Default::default())
 	}
 
@@ -381,12 +420,15 @@ where
 			match self.data.entry(key) {
 				Entry::Occupied(mut entry) => {
 					if entry.get().1 < 0 {
+						self.malloc_tracker.on_insert(&value);
+						self.malloc_tracker.on_remove(&entry.get().0);
 						entry.get_mut().0 = value;
 					}
 
 					entry.get_mut().1 += rc;
 				}
 				Entry::Vacant(entry) => {
+					self.malloc_tracker.on_insert(&value);
 					entry.insert((value, rc));
 				}
 			}
@@ -421,53 +463,30 @@ where
 	}
 }
 
-// `no_std` implementation requires that hasmap
-// is implementated in parity-util-mem, that
-// is currently not the case.
-#[cfg(feature = "std")]
-impl<H, KF, T> MallocSizeOf for MemoryDB<H, KF, T>
+impl<H, KF, T, M> MallocSizeOf for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	H::Out: MallocSizeOf,
 	T: MallocSizeOf,
 	KF: KeyFunction<H>,
 	KF::Key: MallocSizeOf,
+	M: MemTracker<T>,
 {
 	fn size_of(&self, ops: &mut MallocSizeOfOps) -> usize {
-		self.data.size_of(ops)
+		self.data.shallow_size_of(ops)
+			+ self.malloc_tracker.get_size()
 			+ self.null_node_data.size_of(ops)
 			+ self.hashed_null_node.size_of(ops)
 	}
 }
 
-// This is temporary code, we should use
-// `parity-util-mem`, see
-// https://github.com/paritytech/trie/issues/21
-#[cfg(not(feature = "std"))]
-impl<H, KF, T> MallocSizeOf for MemoryDB<H, KF, T>
-where
-	H: KeyHasher,
-	H::Out: MallocSizeOf,
-	T: MallocSizeOf,
-	KF: KeyFunction<H>,
-	KF::Key: MallocSizeOf,
-{
-	fn size_of(&self, ops: &mut MallocSizeOfOps) -> usize {
-		use core::mem::size_of;
-		let mut n = self.data.capacity() * (size_of::<T>() + size_of::<H>() + size_of::<usize>());
-		for (k, v) in self.data.iter() {
-			n += k.size_of(ops) + v.size_of(ops);
-		}
-		n + self.null_node_data.size_of(ops) + self.hashed_null_node.size_of(ops)
-	}
-}
-
-impl<H, KF, T> PlainDB<H::Out, T> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> PlainDB<H::Out, T> for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default + PartialEq<T> + for<'a> From<&'a [u8]> + Clone + Send + Sync,
 	KF: Send + Sync + KeyFunction<H>,
 	KF::Key: Borrow<[u8]> + for <'a> From<&'a [u8]>,
+	M: MemTracker<T> + Send + Sync,
 {
 	fn get(&self, key: &H::Out) -> Option<T> {
 		match self.data.get(key.as_ref()) {
@@ -488,11 +507,14 @@ where
 			Entry::Occupied(mut entry) => {
 				let &mut (ref mut old_value, ref mut rc) = entry.get_mut();
 				if *rc <= 0 {
+					self.malloc_tracker.on_insert(&value);
+					self.malloc_tracker.on_remove(old_value);
 					*old_value = value;
 				}
 				*rc += 1;
 			},
 			Entry::Vacant(entry) => {
+				self.malloc_tracker.on_insert(&value);
 				entry.insert((value, 1));
 			},
 		}
@@ -505,28 +527,32 @@ where
 				*rc -= 1;
 			},
 			Entry::Vacant(entry) => {
-				entry.insert((T::default(), -1));
+				let value = T::default();
+				self.malloc_tracker.on_insert(&value);
+				entry.insert((value, -1));
 			},
 		}
 	}
 }
 
-impl<H, KF, T> PlainDBRef<H::Out, T> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> PlainDBRef<H::Out, T> for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default + PartialEq<T> + for<'a> From<&'a [u8]> + Clone + Send + Sync,
 	KF: Send + Sync + KeyFunction<H>,
 	KF::Key: Borrow<[u8]> + for <'a> From<&'a [u8]>,
+	M: MemTracker<T> + Send + Sync,
 {
 	fn get(&self, key: &H::Out) -> Option<T> { PlainDB::get(self, key) }
 	fn contains(&self, key: &H::Out) -> bool { PlainDB::contains(self, key) }
 }
 
-impl<H, KF, T> HashDB<H, T> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> HashDB<H, T> for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default + PartialEq<T> + for<'a> From<&'a [u8]> + Clone + Send + Sync,
 	KF: Send + Sync + KeyFunction<H>,
+	M: MemTracker<T> + Send + Sync,
 {
 	fn get(&self, key: &H::Out, prefix: Prefix) -> Option<T> {
 		if key == &self.hashed_null_node {
@@ -562,11 +588,14 @@ where
 			Entry::Occupied(mut entry) => {
 				let &mut (ref mut old_value, ref mut rc) = entry.get_mut();
 				if *rc <= 0 {
+					self.malloc_tracker.on_insert(&value);
+					self.malloc_tracker.on_remove(old_value);
 					*old_value = value;
 				}
 				*rc += 1;
 			},
 			Entry::Vacant(entry) => {
+				self.malloc_tracker.on_insert(&value);
 				entry.insert((value, 1));
 			},
 		}
@@ -594,38 +623,43 @@ where
 				*rc -= 1;
 			},
 			Entry::Vacant(entry) => {
-				entry.insert((T::default(), -1));
+				let value = T::default();
+				self.malloc_tracker.on_insert(&value);
+				entry.insert((value, -1));
 			},
 		}
 	}
 }
 
-impl<H, KF, T> HashDBRef<H, T> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> HashDBRef<H, T> for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default + PartialEq<T> + for<'a> From<&'a [u8]> + Clone + Send + Sync,
-	KF: Send + Sync + KeyFunction<H>,
+	KF: KeyFunction<H> + Send + Sync,
+	M: MemTracker<T> + Send + Sync,
 {
 	fn get(&self, key: &H::Out, prefix: Prefix) -> Option<T> { HashDB::get(self, key, prefix) }
 	fn contains(&self, key: &H::Out, prefix: Prefix) -> bool { HashDB::contains(self, key, prefix) }
 }
 
-impl<H, KF, T> AsPlainDB<H::Out, T> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> AsPlainDB<H::Out, T> for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default + PartialEq<T> + for<'a> From<&'a[u8]> + Clone + Send + Sync,
-	KF: Send + Sync + KeyFunction<H>,
+	KF: KeyFunction<H> + Send + Sync,
 	KF::Key: Borrow<[u8]> + for <'a> From<&'a [u8]>,
+	M: MemTracker<T> + Send + Sync,
 {
 	fn as_plain_db(&self) -> &dyn PlainDB<H::Out, T> { self }
 	fn as_plain_db_mut(&mut self) -> &mut dyn PlainDB<H::Out, T> { self }
 }
 
-impl<H, KF, T> AsHashDB<H, T> for MemoryDB<H, KF, T>
+impl<H, KF, T, M> AsHashDB<H, T> for MemoryDB<H, KF, T, M>
 where
 	H: KeyHasher,
 	T: Default + PartialEq<T> + for<'a> From<&'a[u8]> + Clone + Send + Sync,
-	KF: Send + Sync + KeyFunction<H>,
+	KF: KeyFunction<H> + Send + Sync,
+	M: MemTracker<T> + Send + Sync,
 {
 	fn as_hash_db(&self) -> &dyn HashDB<H, T> { self }
 	fn as_hash_db_mut(&mut self) -> &mut dyn HashDB<H, T> { self }
@@ -634,6 +668,7 @@ where
 #[cfg(test)]
 mod tests {
 	use super::{MemoryDB, HashDB, KeyHasher, HashKey};
+	use parity_util_mem::malloc_size;
 	use hash_db::EMPTY_PREFIX;
 	use keccak_hasher::KeccakHasher;
 
@@ -697,5 +732,19 @@ mod tests {
 		let (db2, root) = MemoryDB::<KeccakHasher, HashKey<_>, Vec<u8>>::default_with_root();
 		assert!(db2.contains(&root, EMPTY_PREFIX));
 		assert!(db.contains(&root, EMPTY_PREFIX));
+	}
+
+	#[test]
+	fn malloc_size_of() {
+		let mut db = MemoryDB::<KeccakHasher, HashKey<_>, Vec<u8>>::default();
+		for i in 0u32..1024 {
+			let bytes = i.to_be_bytes();
+			let prefix = (bytes.as_ref(), None);
+			db.insert(prefix, &bytes);
+		}
+		assert_eq!(
+			malloc_size(&db),
+			malloc_size(&db.data) + malloc_size(&db.null_node_data) + malloc_size(&db.hashed_null_node)
+		);
 	}
 }