+++ memory-db/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog].
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
+- Type parameter to count malloc_size_of on memory-db [#94](https://github.com/paritytech/trie/pull/94)
 
 ## [0.20.0] - 2020-03-21
 - Update parity-util-mem to v0.6 [#82](https://github.com/paritytech/trie/pull/82)